Cleanup code migration to wxWidgets v3.0
- Resolve issue regarding handling button events. (toolbar crash issue)
- Additional tracing for of Toolbar crash issue. (to be removed later)
- Remove depreciated calls to libpng misc warning cleanup.
- Resolve debug print format issue in StringBase.
- Correct error in GetImageTypeFromFileName.
- Misc cleanup of asserts warning associated with wx3.0.
- Eliminate depreciated parameter in slider method.
- Resolve image loading issue.
- Tweak build script.
- Tweak ftfonts.cpp use unique name for local wxGetRootWindow.

+++ wxOil/makemsg.cpp
@@ -1,7 +1,7 @@
 // $Id: makemsg.cpp 1333 2006-06-16 20:34:16Z alex $
 /* @@tag:xara-cn@@ DO NOT MODIFY THIS LINE
 ================================XARAHEADERSTART===========================
- 
+
                Xara LX, a vector drawing and manipulation program.
                     Copyright (C) 1993-2006 Xara Group Ltd.
        Copyright on certain contributions may be held in joint with their
@@ -32,7 +32,7 @@ ADDITIONAL RIGHTS
 
 Conditional upon your continuing compliance with the GNU General Public
 License described above, Xara Group Ltd grants to you certain additional
-rights. 
+rights.
 
 The additional rights are to use, modify, and distribute the software
 together with the wxWidgets library, the wxXtra library, and the "CDraw"
@@ -96,10 +96,10 @@ service marks of Xara Group Ltd. All rights in these marks are reserved.
 =================================XARAHEADEREND============================
  */
 /*************************************************************************************
-	
-	
+
+
 	Author: Justin_Flude (Xara Group Ltd) <camelotdev@xara.com>
-	
+
 	Definition of the member function StringBase::MakeMsg(), which extends the traditional
 	sprintf() function to cope with arbitrarily ordered format specifications.
 **************************************************************************************/
@@ -138,7 +138,7 @@ INT32 __cdecl StringBase::MakeMsg(UINT32 resID ...)
 
 	va_list ap;
 	va_start(ap, resID);
-	
+
 	INT32 n = 0;
 	StringBase s;
 	if (s.Alloc(MAX_STRING_RES_LENGTH) && s.Load(resID))
@@ -152,21 +152,21 @@ INT32 __cdecl StringBase::MakeMsg(UINT32 resID ...)
 			TRACE( _T("Failed to allocate or load the format string in MakeMsg()\n"));
 	} */
 #endif
-	
+
 	va_end(ap);
-	
+
 #if !defined(EXCLUDE_FROM_RALPH) && !defined(EXCLUDE_FROM_XARALX)
 	// Remember the resource ID, in case the help system can't work out the message ID.
 	SetNextMsgHelpContext(resID);
 #endif
-	
+
 	// Return the number of characters in the formatted string.
 	return n;
 }
 
 
-	
-	
+
+
 /**************************************************************************************
 >	INT32 __cdecl StringBase::_MakeMsg(const TCHAR* fmt ...)
 
@@ -178,7 +178,7 @@ INT32 __cdecl StringBase::MakeMsg(UINT32 resID ...)
 	Outputs:	Changes "this" to become the formatted string, with the passed parameters
 				substituted into the string resource in the correct order.
 	Returns:	The number of characters in the formatted string.
-	Purpose:	Internationally portable version of sprintf(...), eg.					
+	Purpose:	Internationally portable version of sprintf(...), eg.
 					UINT32 x;
 					s._MakeMsg("#1%lu", x);
 					TextOut(hdc, 20, 20, s, s.Length());
@@ -190,14 +190,14 @@ INT32 __cdecl StringBase::_MakeMsg(const TCHAR* fmt ...)
 */
 	ENSURE(text, "Call to StringBase::_MakeMsg for an unallocated String");
 	ENSURE(fmt, "Call to StringBase::_MakeMsg with a null format parameter");
-	
+
 	va_list ap;
 	va_start(ap, fmt);
-	
+
 	INT32 n = CCvsprintf(fmt, ap);
-	
+
 	va_end(ap);
-	return n;	
+	return n;
 }
 
 
@@ -226,15 +226,15 @@ struct Item
 	ArgType type;		// type of associated parameter
 	INT32 pos;			// relative position in format string
 	Item* next;			// next in list
-		
+
 	static Item* head;
 	static Item* tail;
-	
+
 	Item(const TCHAR* s, INT32 ps, ArgType t);
 	~Item();
 };
-	
-	
+
+
 Item* Item::head = 0;
 Item* Item::tail = 0;
 
@@ -251,8 +251,8 @@ Item::Item(const TCHAR* s, INT32 ps, ArgType t)
 }
 
 
-	
-	
+
+
 // Destroy an Item - tail recursively deletes every Item, if called by
 // delete Item::head;
 Item::~Item()
@@ -263,8 +263,8 @@ Item::~Item()
 }
 
 
-	
-	
+
+
 /**************************************************************************************
 >	INT32 StringBase::BuildList(const TCHAR* format)
 
@@ -288,10 +288,10 @@ INT32 StringBase::BuildList(const TCHAR* format)
 */
 	ENSURE(format, "Null parameter passed to StringBase::BuildList");
     const TCHAR* start = format;
-    INT32 nArgs = 0;    
+    INT32 nArgs = 0;
     while (*start)
     {
-    	const TCHAR* next = start;    	
+    	const TCHAR* next = start;
     	// Look for a format specifier in the layout string
     	if (*next == TEXT('#'))
     		if (IsNumeric(*(++next)))
@@ -310,7 +310,7 @@ INT32 StringBase::BuildList(const TCHAR* format)
     				while (IsNumeric(*next)) next++;
     				if (*next == TEXT('.'))
     					while (IsNumeric(*++next));
-    				
+
     				// Ok, we have skipped the width, precision etc. and
 					// "next" should now be pointing at the type character(s)
 					ArgType kind;
@@ -329,7 +329,7 @@ INT32 StringBase::BuildList(const TCHAR* format)
 						}
 					else
 						switch (*next)			// single type character
-						{    
+						{
 							case TEXT('c'):
 								kind = CHAR_ARG;
 								break;
@@ -346,13 +346,13 @@ INT32 StringBase::BuildList(const TCHAR* format)
 								kind = CHAR_POINTER_ARG;
 								break;
 							case TEXT('S'):
-								kind = STRING_POINTER_ARG;								
+								kind = STRING_POINTER_ARG;
 								break;
 							default:
 								next = start + 1;
 								goto not_format;			// boo hiss hooray!
 						}
-					
+
 					// Successfully parsed the format specifier, so add
 					// it to the Item list.  Skip the leading #n of the specifier,
 					// so (next) now points to the character following the '%X'
@@ -367,7 +367,7 @@ INT32 StringBase::BuildList(const TCHAR* format)
 					new Item(temp, ArgPos, kind);
 					start = next;
 					nArgs++;
-					
+
 /*					if (IsUserName("JustinF"))
 					{
 						TRACE( _T("\tItem: (%d) at #%d  %s\n"), kind, ArgPos, temp);
@@ -380,8 +380,8 @@ INT32 StringBase::BuildList(const TCHAR* format)
     				// Put back the previous character
     				next--;
     		}
-    	
-not_format:    	
+
+not_format:
     	// Scan a literal up to the next '#' or null, whichever is sooner
     	while (*next && *next != TEXT('#')) next++;
 
@@ -400,11 +400,11 @@ not_format:
 */
     }
 
-	return nArgs;    	
+	return nArgs;
 }
 
 
-	
+
 
 /**************************************************************************************
 >	INT32 StringBase::CCvsprintf(const TCHAR* layout, va_list va)
@@ -455,7 +455,7 @@ INT32 StringBase::CCvsprintf(const TCHAR* layout, va_list va)
 						break;
 					// type 'ld' / 'li' - signed decimal INT32
 					case SIGNED_INT32_ARG:
-						camSnprintf(temp, 512, p->str, va_arg(va, INT32));
+						camSnprintf(temp, 512, p->str, va_arg(va, signed long));
 						break;
 					// type 'u' - unsigned decimal integer
 					case UNSIGNED_INT_ARG:
@@ -463,7 +463,7 @@ INT32 StringBase::CCvsprintf(const TCHAR* layout, va_list va)
 						break;
 					// type 'lu' / 'lx' / 'lX' - unsigned decimal INT32
 					case UNSIGNED_INT32_ARG:
-						camSnprintf(temp, 512, p->str, va_arg(va, UINT32));
+						camSnprintf(temp, 512, p->str, va_arg(va, unsigned long));
 						break;
 					// type 's' - long pointer to array of (constant) char
 					case CHAR_POINTER_ARG:
@@ -490,7 +490,7 @@ INT32 StringBase::CCvsprintf(const TCHAR* layout, va_list va)
 				break;
 			}
 	}
-	
+
 	// Concatenate every string in the list, tidy up, and return in "this" string
 	*text = 0;
 	UINT32 NextChar = 0;
